Move DocumentTransform to utilities sub-package to avoid circular dependency
